Interactive visualizations (#54)
* Created AppState singleton class. Storing visualization option in AppState.

* Added BasicPlot, ModelBase classes.

BasicPlot is using a PyQt GUI (extremely customizable, general purpose GUI library), that embeds the matplolib stuff. It is intended to be the default plotting system for the models, and just shows input, output, and target sequences.
ModelBase is intended to be the base class of all the models, such that by default the models get the BasicPlot through the `plot_sequence` method. That method is intended to be overriden to make more advanced plots adapted to the model.

* Changed BasicPlot to TimePlot. Plots list of figures.

* Ported DWM model to new time_plot system.